feature: Set up Routes #22
created routes for ThinkingSession (i.e. a session of the thinking hats
process) and Cards (i.e. synonymous to Ideas) and dummy functions in
respective controllers. Hats will not need a dedicated controller as it
is a mere state variable of a session. Still it will be a model object
to handle the UI changes depending on the current hat easier (by using
it as a renderArg)