WebSocket Prototype:
Naming: Each ThinkingSession (designated by thinkingSession_X, with X being id) has several TOPICS.
These topics can be subscribed to and published to by the client. Each client that is subscribed to a topic will be notified if something is published to it.

Each interaction with the app (add card, move card, switch hat etc) should be a topic of its own and all clients should be subscribed to them by default.